Moved db functions into seperate file
